Move navigation out of root directory
