+++ src/panel/App.tsx
@@ -3,7 +3,7 @@ import React, { FC } from "react";
 import { HashRouter, Route, Redirect } from "react-router-dom";
 import { ThemeProvider } from "styled-components";
 import { Disconnected, Explorer, Request, Timeline, Mismatch } from "./pages";
-import { Navigation } from "./Navigation";
+import { Navigation } from "./components/Navigation";
 import { theme, GlobalStyle } from "./theme";
 import {
   DevtoolsProvider,
@@ -37,6 +37,13 @@ export const AppRoutes: FC = () => {
 
   return (
     <HashRouter>
+      <Navigation
+        items={[
+          { link: "/explorer", label: "Explorer" },
+          { link: "/events", label: "Events" },
+          { link: "/request", label: "Request" },
+        ]}
+      />
       <TimelineProvider>
         <Route path="/events" component={Timeline} />
       </TimelineProvider>
@@ -47,7 +54,6 @@ export const AppRoutes: FC = () => {
         <Route path="/explorer" exact component={Explorer} />
       </ExplorerProvider>
       <Route path="/" exact component={() => <Redirect to="/explorer" />} />
-      <Navigation />
     </HashRouter>
   );
 };
