+++ src/panel/components/Navigation.fixture.tsx
