+++ src/panel/components/index.ts
@@ -4,3 +4,4 @@ export * from "./Pane";
 export * from "./Portal";
 export * from "./CodeHighlight";
 export * from "./Collapsible";
+export * from "./Navigation";
